Extract createImportMergeSlice from projectStore (P5 store slice)
