+++ src/store/projectStore.ts
@@ -26,7 +26,7 @@ import {
   type ImportedModelFormat,
 } from '../engine/importedMesh'
 import { clearSTLTransformedGeometryCache } from '../engine/csg'
-import { createImportedFeature, isProfileDegenerate, mergeCamjFolders, uniqueName } from '../import'
+import { isProfileDegenerate, uniqueName } from '../import'
 import {
   type Segment,
   defaultStock,
@@ -116,6 +116,7 @@ import { createTabsSlice } from './slices/tabsSlice'
 import { createBackdropSlice, normalizeBackdrop } from './slices/backdropSlice'
 import { createMachineDefsSlice } from './slices/machineDefsSlice'
 import { createOperationsSlice } from './slices/operationsSlice'
+import { createImportMergeSlice } from './slices/importMergeSlice'
 import { propagateConstraintsOnTranslate, propagateConstraintsOnRotate, rederiveConstraintGeometry, inferSemanticIndices, validateConstraintsOnFeature, solveFeatureTranslation, type ConstraintInput } from '../sketch/constraintSolver'
 import type {
   PendingAddTool,
@@ -2966,6 +2967,11 @@ export const useProjectStore = create<ProjectStore>((rawSet, get) => {
     uniqueFolderName,
     syncFeatureTreeProject,
   }),
+  ...createImportMergeSlice(set, get, {
+    cloneProject,
+    uniqueFolderName,
+    syncFeatureTreeProject,
+  }),
   ...createFeatureSlice(set, get, {
     cloneProject,
     syncFeatureTreeProject,
@@ -3684,149 +3690,6 @@ export const useProjectStore = create<ProjectStore>((rawSet, get) => {
         sketchEditSession: null,
       }
     }),
-
-
-  importShapes: (input) => {
-    const state = get()
-    const sourceShapes = input.shapes.filter((shape) => !isProfileDegenerate(shape.profile))
-    if (sourceShapes.length === 0) {
-      return []
-    }
-
-    // Group shapes by layer name. Null layer (DXF layer "0") → keyed as '0'.
-    const layerGroups = new Map<string, typeof sourceShapes>()
-    for (const shape of sourceShapes) {
-      const key = shape.layerName ?? '0'
-      const existing = layerGroups.get(key)
-      if (existing) {
-        existing.push(shape)
-      } else {
-        layerGroups.set(key, [shape])
-      }
-    }
-
-    const existingFeatureNames = state.project.features.map((f) => f.name)
-    const newFolders: FeatureFolder[] = []
-    const createdFeatures: SketchFeature[] = []
-
-    let nextProjectLike: Project = {
-      ...state.project,
-      features: [...state.project.features],
-      featureFolders: [...state.project.featureFolders],
-    }
-
-    for (const [layerKey, layerShapes] of layerGroups) {
-      const folderDisplayName = layerKey || '0'
-      const folderId = nextUniqueGeneratedId(nextProjectLike, 'fd')
-      const folderName = uniqueFolderName(folderDisplayName, nextProjectLike.featureFolders)
-      const folder: FeatureFolder = { id: folderId, name: folderName, collapsed: false }
-
-      newFolders.push(folder)
-      nextProjectLike = { ...nextProjectLike, featureFolders: [...nextProjectLike.featureFolders, folder] }
-
-      for (const shape of layerShapes) {
-        const featureName = uniqueName(
-          shape.name || folderDisplayName,
-          [...existingFeatureNames, ...createdFeatures.map((f) => f.name)],
-        )
-        // All closed profiles import as 'add'; open profiles as 'line'.
-        const operation: FeatureOperation = shape.profile.closed ? 'add' : 'line'
-        const nextId = nextUniqueGeneratedId(nextProjectLike, 'f')
-        const feature = normalizeFeatureZRange({
-          ...createImportedFeature(shape, state.project, folderId, featureName, operation),
-          id: nextId,
-        })
-
-        createdFeatures.push(feature)
-        nextProjectLike = { ...nextProjectLike, features: [...nextProjectLike.features, feature] }
-      }
-    }
-
-    if (createdFeatures.length === 0) {
-      return []
-    }
-
-    set((s) => {
-      const nextProject = syncFeatureTreeProject({
-        ...s.project,
-        featureFolders: [...s.project.featureFolders, ...newFolders],
-        featureTree: [
-          ...s.project.featureTree,
-          ...newFolders.map((f) => ({ type: 'folder' as const, folderId: f.id })),
-        ],
-        features: [...s.project.features, ...createdFeatures],
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      })
-      const createdIds = createdFeatures.map((f) => f.id)
-      const primaryId = createdIds.at(-1) ?? null
-      const primaryFolderId = newFolders.at(-1)?.id ?? null
-
-      return {
-        project: nextProject,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: primaryId,
-          selectedFeatureIds: createdIds,
-          selectedNode: primaryId
-            ? { type: 'feature', featureId: primaryId }
-            : primaryFolderId
-              ? { type: 'folder', folderId: primaryFolderId }
-              : s.selection.selectedNode,
-          mode: 'feature',
-          activeControl: null,
-        },
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    })
-
-    return createdFeatures.map((f) => f.id)
-  },
-
-  importCamjFolders: (input) => {
-    const state = get()
-    const merge = mergeCamjFolders({
-      currentProject: state.project,
-      sourceProject: input.sourceProject,
-      selectedFolderIds: input.selectedFolderIds,
-      importStock: input.importStock,
-    })
-    if (merge.createdFeatureIds.length === 0 && !merge.stockReplaced) {
-      return []
-    }
-
-    set((s) => {
-      const nextProject = syncFeatureTreeProject(merge.project)
-      const createdIds = merge.createdFeatureIds
-      const primaryId = createdIds.at(-1) ?? null
-      const primaryFolderId = merge.createdFolderIds.at(-1) ?? null
-      return {
-        project: nextProject,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: primaryId,
-          selectedFeatureIds: createdIds,
-          selectedNode: primaryId
-            ? { type: 'feature', featureId: primaryId }
-            : primaryFolderId
-              ? { type: 'folder', folderId: primaryFolderId }
-              : s.selection.selectedNode,
-          mode: 'feature',
-          activeControl: null,
-        },
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    })
-
-    return merge.createdFeatureIds
-  },
 
   moveFeatureControl: (featureId, control, point) =>
     set((s) => {
